requirements.txt
anyio==3.6.2
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.92.0
filelock==3.9.0
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.12.1
idna==3.4
mypy==1.0.1
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
pydantic==1.10.5
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
